pyproject.toml
[build-system]
requires = ["setuptools>=72", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "reflitao"
version = "0.1.0"
description = "AI Thinking Support Telegram Bot — collect mixed-media artifacts into sessions and query LLMs with full context"
readme = "README.md"
license = { text = "MIT" }
authors = [
    { name = "author" }
]
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "python-telegram-bot>=20",
    "openai>=1.0",
    "anthropic>=0.25",
]

[project.urls]
Homepage = "https://github.com/luisarcher/reflitao-ai/"
Repository = "https://github.com/luisarcher/reflitao-ai/"
Issues = "https://github.com/luisarcher/reflitao-ai/issues"

[project.scripts]
reflitao = "reflitao.__main__:main"

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "coverage>=7",
    "ruff>=0.9",
    "ty>=0.0.1a6",
]

[tool.setuptools.packages.find]
include = ["reflitao*"]

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
version_variables = ["reflitao/__init__.py:__version__"]
build_command = "pip install uv && uv build"
branch = "main"
changelog_file = "CHANGELOG.md"
upload_to_vcs_release = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=reflitao --cov-report=term-missing --cov-report=xml"

[tool.coverage.run]
source = ["reflitao"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 80

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ty.environment]
python-version = "3.11"
